task added

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,5 +1,6 @@
 #include "main.h"
 #include <stdio.h>
+#include <inttypes.h>
 
 /**
  * main - fib0nnaci of first 50 numbers
@@ -8,9 +9,9 @@
 
 int main(void)
 {
-	long int immediate_no = 1;
-	long int previous_no = 0;
-	long int sum_of_immediate_and_previous_no = 0;
+	long long int immediate_no = 1;
+	long long int previous_no = 0;
+	long long int sum_of_immediate_and_previous_no = 0;
 	int counter;
 
 	for (counter = 0; counter < 98; counter++)
@@ -18,7 +19,7 @@ int main(void)
 		sum_of_immediate_and_previous_no = immediate_no + previous_no;
 		previous_no = immediate_no;
 		immediate_no = sum_of_immediate_and_previous_no;
-		printf("%ld", sum_of_immediate_and_previous_no);
+		printf("%llu", sum_of_immediate_and_previous_no);
 		if (counter < 97)
 		{
 			printf(",");
